CCLE-3993 - Removed UCLA specific code.
For CCLE, when a syllabus gets deleted, the user is redirected
to the course's site info page. For now, the open-source syllabus
tool will just redirect to the syllabus page (to inform the user
that the syllabus has been successfully deleted, and to give
the user the opportunity to load a new syllabus).